+++ Rixtrema.WebApi/Rixtrema.DAL/Models/Response/PlansResponses.cs
@@ -4,35 +4,35 @@
     {
         public int Id { get; set; }
         
-        public float? Earnings { get; set; }
+        public double? Earnings { get; set; }
         
-        public float? ParticipantContribAmt { get; set; }
+        public double? ParticipantContribAmt { get; set; }
         
-        public float? EmplrContribIncomeAmt { get; set; }
+        public double? EmplrContribIncomeAmt { get; set; }
         
-        public float? ParticswithBal { get; set; }
+        public double? ParticswithBal { get; set; }
         
-        public float? ActPartics { get; set; }
+        public double? ActPartics { get; set; }
         
-        public float? PartRate { get; set; }
+        public double? PartRate { get; set; }
         
-        public float? AvgPartContrib { get; set; }
+        public double? AvgPartContrib { get; set; }
         
-        public float? AvgEmpContrib { get; set; }
+        public double? AvgEmpContrib { get; set; }
         
-        public float? Assets { get; set; }
+        public double? Assets { get; set; }
         
-        public float? Adminexp { get; set; }
+        public double? Adminexp { get; set; }
         
-        public float? AvgBalance { get; set; }
+        public double? AvgBalance { get; set; }
         
-        public float? PartContribRate { get; set; }
+        public double? PartContribRate { get; set; }
         
-        public float? EmpContribIncomeRate { get; set; }
+        public double? EmpContribIncomeRate { get; set; }
         
-        public float? AdminExpRate { get; set; }
+        public double? AdminExpRate { get; set; }
         
-        public float? PercRetirees { get; set; }
+        public double? PercRetirees { get; set; }
         
         public string SponsDfeMailState { get; set; }
